1. Improved usage of custom exceptions 2. Renamed async managers AsyncOpenModules -> AsyncLoadModules, AsyncOpenBooks -> AsyncOpenModule 3. Call myLibararian.loadModules() with myLibararian initialization 4. The start home activity (Books or Reader) depends on last_read settings. (see BibleQuote)
