--- src/com/BibleQuote/exceptions/CreateModuleErrorException.java
@@ -1,33 +1,0 @@
-/*
- * Copyright (C) 2011 Scripture Software (http://scripturesoftware.org/)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.BibleQuote.exceptions;
-
-public class CreateModuleErrorException extends Exception {
-
-	private static final long serialVersionUID = -1652902166548627455L;
-
-	public String toString(){
-		return "CreateModuleErrorException";
-	}
-	
-	public CreateModuleErrorException(String message) {
-		super(message);
-	}
-
-	public CreateModuleErrorException(Exception parent) {
-		super(parent);
-	}
-}
